Fix the BeatSensitivity type, and one style catalog instead of three
BeatSensitivity.minIntervalMs was typed Int; the refractory gate takes a
Float all the way through (GuiPrefs.beatMinIntervalMs ->
FeatureTimeline.withBeatSensitivity), so the previous commit did not
compile. Corrected to Float — that was the only defect in it.

Then a real bug the mapping turned up, and the fix that prevents its
whole class. "Which styles exist" was written out three times:
VisualizerRenderer.availableSceneIds, the Styles tab in VisualsHub, and
PlayerViewModel.randomStepNow. The three had drifted — the random pool
was `PARTICLE_SCENES + SHADER_SCENES.keys` only, so with Random visuals
and "include styles" on, Fluid, Curl Flow and Water could never come up.
They were reachable only sideways, via the built-in presets that happen
to use them. Nothing in the code marked this as deliberate.

SceneCatalog now owns the list and all three read it. randomStyles() is
defined as selectableStyles() minus MilkDrop, so "anything you can pick by
hand, random can pick too" is expressed rather than re-typed, and
SceneCatalogTest pins it. The particle and shader id lists stay on
VisualizerRenderer for now — they carry its shader resource ids, and
folding them in is the wider scene-registry step.

AudioQualityTrackerTest pins the seam the previous round extracted: the
readout is a join of two events from different threads (either can arrive
first), and the container is read from the uri at event time.

+++ musicviz-project/musicviz/app/src/main/java/dev/musicviz/ui/ExportCoordinator.kt
@@ -16,10 +16,16 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-/** Beat-detection settings in force right now, applied to every export. */
+/**
+ * Beat-detection settings in force right now, applied to every export.
+ *
+ * [minIntervalMs] is a Float because that is what the refractory gate takes —
+ * it comes off a slider and feeds [FeatureTimeline.withBeatSensitivity]
+ * directly.
+ */
 data class BeatSensitivity(
     val thresholdSigma: Float,
-    val minIntervalMs: Int,
+    val minIntervalMs: Float,
 )
 
 /** The visual configuration an export renders with, sampled when it starts. */
